refactor Counter
