+++ jsconfig.json
@@ -1,5 +1,6 @@
 {
   "compilerOptions": {
     "baseUrl": "src"
-  }
+  },
+  "include": ["src"]
 }
