+++ src/features/counter/index.js
@@ -1,3 +1,3 @@
 export {default as CounterReducer} from './CounterReducer';
 export {default as useCountValue} from './selectors';
-export {default as useActions} from './actionCreators';
+export {default as useIncrementCounter} from './useIncrementCounter';
