+++ src/features/counter/useIncrementCounter.js
@@ -3,16 +3,15 @@ import {useDispatch} from 'react-redux';
 import useCountValue from './selectors';
 import {INCREMENT_COUNTER} from './actionTypes';
 
-const useActions = () => {
+const useIncrementCounter = () => {
   const dispatch = useDispatch();
   const count = useCountValue();
-  const incrementCounter = useCallback(() => {
+  return useCallback(() => {
     dispatch({
       type: INCREMENT_COUNTER,
       value: count + 1,
     });
   }, [count, dispatch]);
-  return {incrementCounter};
 };
 
-export default useActions;
+export default useIncrementCounter;
